Reorganize path and common variables.
